repipe: Update submodule pointers for libreconf and libusrxml
Update submodule pointer for libreconf to 8d9e3fff861b.
Update submodule pointer for libusrxml to c52f1ae0d9fd.

With this submodule update prepare_file() function improved to accept
either absolute path to file to prepare that is relative to ROOT or
relative path that is prefixed with ROOT internally.

Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>